fix(gateway): break ws-connection import cycle, satisfy typecheck
