+++ src/gateway/server/ws-connection/message-handler.post-connect-health.test.ts
@@ -137,6 +137,7 @@ describe("attachGatewayWsMessageHandler post-connect health refresh", () => {
         return true;
       },
       setHandshakeState: vi.fn(),
+      advanceHandshakePhase: vi.fn(),
       setCloseCause: vi.fn(),
       setLastFrameMeta: vi.fn(),
       originCheckMetrics: { hostHeaderFallbackAccepted: 0 },
